refactor(crypto): move verifySignatures into Transactions.Verifier
